requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
pydantic
python-dotenv
email-validator
